-Move to localhost but using Raspi ip first -Integrate null pointer handling for serveral objects
