+++ RaspberryDebugger/RaspberryDebuggerPackage.cs
@@ -32,7 +32,7 @@ using Microsoft.VisualStudio.Shell.Interop;
 
 using Neon.Common;
 using Neon.Diagnostics;
-
+using RaspberryDebugger.Commands;
 using Task = System.Threading.Tasks.Task;
 
 namespace RaspberryDebugger
